% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condprob.R
\name{condprob}
\alias{condprob}
\title{Function condprob}
\usage{
condprob(xX, xY, xImpair, ProbComp = c("gt", "lt"), Exceed = c("gte",
  "lte"), ci = FALSE, alpha = 0.05, R = 100, xW = rep(1, length(xX)))
}
\arguments{
\item{xW}{}
}
\description{
This function calculates a conditional probability analysis
}
\examples{
\dontrun{
data(jeqdata_wq)
ept<-jeqdata_wq$EPT.RICH
pct_fn<-jeqdata_wq$PCT.FN
cpa1<-condprob(pct_fn,ept,9,'lt','gte',T,R=1000)
}
}

