% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conmat_to_df.R
\name{conmat_to_df}
\alias{conmat_to_df}
\title{convert table to df}
\usage{
conmat_to_df(cm_tab)
}
\arguments{
\item{cm_tab}{a table class with confusion matrix}
}
\description{
convert table to df
}

