% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chla.R
\name{get_chla}
\alias{get_chla}
\title{get_chla}
\usage{
get_chla(cp_obj, prob)
}
\arguments{
\item{cp_obj}{An object of class condprob is required.}

\item{prob}{Exceedence probability to return chl a}
}
\description{
Returns Chl a value that is associated with a given probability of exceedence at the
Upper bootstrapped confidence interval
}

