% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nla.R
\name{get_nla}
\alias{get_nla}
\title{get_nla}
\usage{
get_nla(data_url)
}
\arguments{
\item{data_url}{URL pointing to the National Lakes Assessment data}
}
\description{
Get data from US EPA NLA site and process based on a few common filters
}
\examples{
nla_dat <- get_nla("http://water.epa.gov/type/lakes/assessmonitor/lakessurvey/upload/NLA2007_Recreational_ConditionEstimates_20091123.csv")
}

