% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_cdf.R
\name{plot_cdf}
\alias{plot_cdf}
\title{plotCDF Function to create CDFs of chla and microcystin}
\usage{
plot_cdf(df, ...)
}
\arguments{
\item{df}{a data frame with all parameters to plot}
}
\description{
This functions produces CDF plots for the Hollister et al. microcystin - chla paper
}

