% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{get_nla}
\usage{
plot_cp(cp, ...)
}
\arguments{
\item{cp}{A conditional probability object from condprob2}

\item{...}{arguments for labels}
}
\description{
Create plot of conditional probability object showing association between microcystin-LR
advisory levels and chlorophyll \italics{a}
}

